Organize commands and options into domain folders
Mirror the Core/Tools structure (Pipeline/, Workspace/) in the Fabric
adapter project for commands and options.

Structure:
  Commands/Pipeline/ - ListPipelines, CreatePipeline, GetPipeline, RunPipeline
  Commands/Workspace/ - ListWorkspaces
  Options/Pipeline/ - Pipeline-specific options
  Options/Workspace/ - Workspace-specific options
  Options/ - Shared option definitions
