+++ DataFactory.MCP.Fabric/Commands/Pipeline/ListPipelinesCommand.cs
@@ -3,6 +3,7 @@
 
 using DataFactory.MCP.Fabric.Models;
 using DataFactory.MCP.Fabric.Options;
+using DataFactory.MCP.Fabric.Options.Pipeline;
 using DataFactory.MCP.Handlers.Pipeline;
 using Microsoft.Extensions.Logging;
 using Microsoft.Mcp.Core.Commands;
@@ -11,7 +12,7 @@ using Microsoft.Mcp.Core.Models.Command;
 using Microsoft.Mcp.Core.Models.Option;
 using Microsoft.Mcp.Core.Options;
 
-namespace DataFactory.MCP.Fabric.Commands;
+namespace DataFactory.MCP.Fabric.Commands.Pipeline;
 
 [CommandMetadata(
     Id = "a1b2c3d4-1001-4000-8000-000000000002",
