+++ DataFactory.MCP.Fabric/Options/Workspace/ListWorkspacesOptions.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.Mcp.Core.Options;
 
-namespace DataFactory.MCP.Fabric.Options;
+namespace DataFactory.MCP.Fabric.Options.Workspace;
 
 public sealed class ListWorkspacesOptions : GlobalOptions
 {
